Remove use of shared file
