+++ src/_includes/foot.html
@@ -40,7 +40,6 @@
   {% endif %}
 
   {% if site.data.toggles.slt-397-website-header-updates == 'on' %}
-    <script src="{{  '/assets/custom/js/websiteHeaderShared.js' | prepend: site.baseurl  }}"></script>
     <script src="{{ '/assets/custom/js/websiteHeader.js' | prepend: site.baseurl }}"></script>
     <script src="{{ '/assets/custom/js/websiteNavigation.js' | prepend: site.baseurl }}"></script>
   {% endif %}
