+++ src/assets/custom/js/websiteHeader.test.js
@@ -1,5 +1,4 @@
 const { simulatePageLoad } = require("./simulatePageLoad");
-require("./websiteHeaderShared");
 require("./websiteHeader");
 
 let header;
